Move feature to Wasabi.Fluent
