+++ WalletWasabi.Fluent/AppServices/Tor/IIssueListParser.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace WalletWasabi.Tor.NetworkChecker;
+namespace WalletWasabi.Fluent.AppServices.Tor;
 
 public interface IIssueListParser
 {
