+++ WalletWasabi.Fluent/ViewModels/StatusIcon/StatusIconDesignViewModel.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.Windows.Input;
 using ReactiveUI;
-using WalletWasabi.Tor.NetworkChecker;
+using WalletWasabi.Fluent.AppServices.Tor;
 
 namespace WalletWasabi.Fluent.ViewModels.StatusIcon;
