+++ WalletWasabi.Fluent/Views/StatusIcon/TorIssuesControl.axaml
@@ -4,6 +4,7 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:statusIcon="clr-namespace:WalletWasabi.Fluent.ViewModels.StatusIcon"
              xmlns:tnc="clr-namespace:WalletWasabi.Tor.NetworkChecker;assembly=WalletWasabi"
+             xmlns:tor="clr-namespace:WalletWasabi.Fluent.AppServices.Tor"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
              x:Class="WalletWasabi.Fluent.Views.StatusIcon.TorIssuesControl"
              x:CompileBindings="True"
@@ -26,7 +27,7 @@
           <TextBlock DockPanel.Dock="Bottom" Text="Click this icon for details" />
           <ItemsControl Margin="8" Items="{Binding TorIssues}">
             <ItemsControl.ItemTemplate>
-              <DataTemplate x:DataType="tnc:Issue">
+              <DataTemplate x:DataType="tor:Issue">
                 <TextBlock Text="{Binding Title, StringFormat='{}· {0}'}" />
               </DataTemplate>
             </ItemsControl.ItemTemplate>
