+++ WalletWasabi.Fluent/WalletWasabi.Fluent.csproj
@@ -27,7 +27,6 @@
 		<PackageReference Include="Avalonia.Xaml.Behaviors" Version="0.10.13.1" />
 		<PackageReference Include="Avalonia.Skia" Version="$(AvaloniaVersion)" />
 		<PackageReference Include="System.Drawing.Common" Version="6.0.0" />
-		<PackageReference Include="System.Reactive" Version="5.0.0" />
 		<PackageReference Include="System.Runtime" Version="4.3.1" />
 	</ItemGroup>
 	<ItemGroup>
